final batch of issues fixed
