+++ extensions/terminal-suggest/src/completions/code-tunnel-insiders.ts
@@ -2,8 +2,8 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { commonOptions, extensionManagementOptions, troubleshootingOptions, globalTunnelOptions, codeTunnelSubcommands } from './code';
-import codeTunnelCompletionSpec, { codeTunnelOptions, extTunnelSubcommand } from './code-tunnel';
+import { commonOptions, extensionManagementOptions, troubleshootingOptions, globalTunnelOptions, codeTunnelSubcommands, extTunnelSubcommand, codeTunnelOptions } from './code';
+import codeTunnelCompletionSpec, { codeTunnelSpecOptions } from './code-tunnel';
 
 const codeTunnelInsidersCompletionSpec: Fig.Spec = {
 	...codeTunnelCompletionSpec,
@@ -15,7 +15,8 @@ const codeTunnelInsidersCompletionSpec: Fig.Spec = {
 		...extensionManagementOptions('code-tunnel-insiders'),
 		...troubleshootingOptions('code-tunnel-insiders'),
 		...globalTunnelOptions,
-		...codeTunnelOptions
+		...codeTunnelOptions,
+		...codeTunnelSpecOptions
 	]
 };
